requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
customtkinter==5.1.2
darkdetect==0.8.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
openai==0.27.4
ping3==4.0.4
python-dotenv==1.0.0
pywin32==306
requests==2.28.2
Send2Trash==1.8.2
tqdm==4.65.0
urllib3==1.26.15
winshell==0.6
yarl==1.8.2
